requirements.txt
torch>=2.3
torchvision>=0.18.0
transformers>=4.47.0
datasets>=3.1.0
accelerate>=1.1.0
peft>=0.13.0
pillow>=10.0.0
streamlit>=1.40.0
pandas>=2.2.0
numpy>=1.26.0,<2
safetensors>=0.4.5
